backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
